Split text and token classification into separate modules
* Naming follows `transformers.pipeline`.
