+++ setup.cfg
@@ -36,8 +36,8 @@ install_requires =
 
 [options.entry_points]
 spacy_factories =
-    ext_txt_cls_trf = spacy_transformers_pipeline.pipeline_components:make_txt_cls_trf
-    ext_tok_cls_trf = spacy_transformers_pipeline.pipeline_components:make_tok_cls_trf
+    ext_txt_cls_trf = spacy_transformers_pipeline.text_classification:make_txt_cls_trf
+    ext_tok_cls_trf = spacy_transformers_pipeline.token_classification:make_tok_cls_trf
 
 [bdist_wheel]
 universal = true
